requirements.txt
faiss_cpu==1.7.2
ndjson==0.3.1
numpy==1.23.0
openai==0.20.0
tqdm==4.64.0
boto3
Flask
gunicorn
python-dotenv
